Atom feed should have absolute uris
I didn't actually check out this code to test it (I don't have a working vm with Mirage right now) but I'm a lurker and the relative links in the Atom feed were starting to bug me. I'm open to other ways to accomplish this if there's a better one.