Brute-force port of old mirage-www to new style.
This leads to horrendous code style, since we pass read callbacks
everywhere, but this code is about to be retired in favour of a
template KV_RO anyway.  This is just to get it compiling and serving
traffic with 1.0.  Note that this port is also slower than the old
code since the HTML is no longer memoized (but calculated on every
request).  Again, not a big deal since this will be addressed in the
Cowabloga template system.