Remove "doing nothing with functors" sections
Despite the title, this section does not give any example of how to
define or use functors. IMO, it is rather likely to confuse readers,
since it speaks of functors and dependencies, but doesn't give any
examples of either functors or how to use them to manage dependencies!
In the next commit, I integrate the relevant explanatory parts with the
example that actually shows how to use functors for dependency
management.